src/third_party/blink/web_tests/external/wpt/fetch/api/basic/request-upload.any-expected.txt

This is a testharness.js-based test.
PASS Fetch with PUT with body
PASS Fetch with POST with text body
PASS Fetch with POST with URLSearchParams body
PASS Fetch with POST with Blob body
PASS Fetch with POST with ArrayBuffer body
PASS Fetch with POST with Uint8Array body
PASS Fetch with POST with Int8Array body
PASS Fetch with POST with Float32Array body
PASS Fetch with POST with Float64Array body
PASS Fetch with POST with DataView body
PASS Fetch with POST with Blob body with mime type
FAIL Fetch with POST with ReadableStream assert_equals: expected "Test" but got "[object ReadableStream]"
FAIL Fetch with POST with ReadableStream containing String assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL Fetch with POST with ReadableStream containing null assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL Fetch with POST with ReadableStream containing number assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL Fetch with POST with ReadableStream containing ArrayBuffer assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL Fetch with POST with ReadableStream containing Blob assert_unreached: Should have rejected: undefined Reached unreachable code
Harness: the test ran to completion.
